requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.3
cymem==2.0.11
Django==5.2.8
docx2txt==0.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
murmurhash==1.0.13
numpy==2.3.4
packaging==25.0
pdfminer.six==20251107
preshed==3.0.10
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
requests==2.32.5
setuptools==80.9.0
smart_open==7.5.0
spacy==3.8.8
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
thinc==8.3.8
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.2
wrapt==2.0.1
